+++ src/components/assets/transfer/styles/local-transfer.scss
@@ -37,6 +37,10 @@
   margin-top: 32px;
 }
 
+.box--margin-adjuster {
+  margin-top: 8px;
+}
+
 .row--box-error {
   padding: 10px 20px;
   border-radius: 6px;
